+++ example/example.dart
@@ -5,8 +5,8 @@
 import 'package:shelf/shelf_io.dart' as shelf_io;
 import 'package:shelf_proxy/shelf_proxy.dart';
 
-void main() async {
-  var server = await shelf_io.serve(
+Future<void> main() async {
+  final server = await shelf_io.serve(
     proxyHandler('https://dart.dev'),
     'localhost',
     8080,