+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.0.0
+
+- Require Dart `2.12`.
+- Enable null safety.
+- Removed deprecated `createProxyHandler`.
+
 ## 0.1.0+7
 
 * Added example.